requirements.txt
openrouter
pandas
python-dotenv
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
